Backend /my-django-backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
python-dotenv>=0.19,<1.0
psycopg2-binary>=2.9,<3.0
django-cors-headers>=3.10,<4.0
